renamed genre image module to avoid confusion
